Rename package to spork
