+++ src/main/java/org/kjy5/spork/ClassRepresentatives.java
@@ -1,7 +1,7 @@
 /*
  * Copyright (c) 2024 Kenneth Yang (kjy5@uw.edu)
  */
-package org.kjy5.tdm;
+package org.kjy5.spork;
 
 import com.github.gumtreediff.matchers.MappingStore;
 import com.github.gumtreediff.tree.Tree;
